requirements.txt
streamlit
torch
transformers
faiss-cpu
